added inputaddress component to input component
