+++ src/components/ui/Input.tsx
@@ -19,7 +19,6 @@ function Input({
   onChange: React.ChangeEventHandler<HTMLInputElement> | undefined,
   onKeyDown: React.KeyboardEventHandler<HTMLInputElement> | undefined,
 }) {
-  // border-black-100
   return (
     <input
       type={type}
@@ -38,6 +37,35 @@ function Input({
   );
 }
 
+const InputAddress = ({
+  value,
+  disabled,
+  placeholder,
+  error,
+  onChange,
+}: {
+  value: string,
+  disabled: boolean,
+  placeholder: string,
+  error: boolean,
+  onChange: (newValue: string) => void,
+}) => {
+return (
+  <Input
+  width="w-full"
+  value={value}
+  type="text"
+  min={undefined}
+  disabled={disabled}
+  placeholder={placeholder}
+  borderColor={error ? "border-red" : "border-black-100"}
+  onChange={e => onChange(e.target.value)}
+  onKeyDown={undefined}
+/>
+)
+}
+
 export {
   Input,
+  InputAddress
 };
